refactoring classes directory
