+++ classes/common/timer.js
