+++ classes/trivia/question.js
@@ -1,8 +1,8 @@
 const { EmbedBuilder }  = require('discord.js');
 const HE = require('he');
 const Sequelize = require('sequelize');
-const { Answer } = require('./trivia/answer.js');
-const { Timer } = require('./timer.js');
+const { Answer } = require('./answer.js');
+const { Timer } = require('../common/timer.js');
 const Reactions = [
     '\u0031\u20E3',     // :one: 
     '\u0032\u20E3',     // :two:
